Add info fragment, viewbinding, transport list adapter cleanup

+++ app/src/main/java/hu/selester/selnet/database/tables/TasksTable.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 
 @Entity
 data class TasksTable(
-    @PrimaryKey(autoGenerate = true) var Id: Int?,
+    @PrimaryKey(autoGenerate = true) var Id: Long?,
     var orderId: Long,
     var company: String,
     var shortInfo: String,
